requirements.txt
Click==7.0
Flask-Bootstrap==3.3.0.1
Flask==1.0.2
Jinja2==2.10.1
MarkupSafe==1.1.1
Mock==1.0.1
Werkzeug==0.15.2
attrs==17.4.0
autopep8==1.3.3
croniter==0.3.10
funcsigs==1.0.2
itsdangerous==1.1.0
pluggy==0.6.0
py==1.5.2
pycodestyle==2.3.1
python-crontab==1.9.2
six==1.11.0
pytest==5.2.1
pytest-mock==1.11.1
